backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Mem0 Memory System
mem0ai==0.0.10

# Google Cloud AI APIs
google-generativeai==0.3.1
google-ai-python==0.1.0

# Vector Database
chromadb==0.4.18

# Database
sqlite3  # Built-in, listed for clarity

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
pytest-benchmark==4.0.0

# Code Quality
ruff==0.1.7
black==23.11.0

# Utilities
python-multipart==0.0.6
typing-extensions==4.8.0
